Reverted slurs to what they were before
We'd have liked to implement:

Slur (List Music) where we can hold all the music contained by the slur and balance the brackets.  Unfortunately we revert to Slur Char and just recognize the bracket characters that frame the slurred notes.  This is because in the wild, there are ABC examples where the slur extends across music lines which makes the preferred implementation impossible.