Added a utility to handle some music notation
What you get back from the parser is not always in exactly the format you'd like. For example, a note in the tune body may or may not have an explicit accidental associated with it.  If it does not, it may still have an implicit accidental associated because of the key signature.  (for example, an F in the key of G has an implied sharp).  It is useful to have a few library functions to create the correct musical object (e.g. F#) from the parse tree.