Made bar lines type-safe
Bar line options are extremely badly specified in the spec (because of backward compatibility issues to previous spec versions that were too loosely specified). Defined a Repeat disjunction data type to try to encapsulate sensible layout options for repeated sections, also a line count,  and altered the Bar type accordingly.  Interpreted this as sensibly as possible in Canonical ABC.

Also, removed the 'parse any crud at the end of the ABC' option that was there for debug purposes and is now unnecessary with the parser becoming more mature.