Upgraded to Combine 2.02
Also added custom versions of the manyTill and or combinators to allow different forms of error reporting.  Sorted out the top-level syntax a bit so that manyTill can be used consistently.  All this together means we can point accurately to individual errors within an ABC line rather than just to the start of the line.