Influence of explicit accidentals in bars
If a note is marked explicitly as an accidental then all succeeding notes within the same bar inherit the accidental  (i.e. you only need to mark the first occurrence in any bar).

Causes a major change to AbcPerformance.  We had been doing fold right everywhere but this now forces uss to process rigorously left to right - so we fold left everywhere and then reverse the result.

Notation altered so that the accidental behaviour is ordered:

1) Explicit accidental on the note itself
2) Explicit accidental on the same note earlier in the bar
3) Implicit accidental inherited from the key signature