Support for transposition amidst key changes
Normal transposition works out a key distance between the source and the target keys.  If the tune changes key with an inline K header, then this is accommodated by working out a new target key by adding the transposition distance.

This works OK, but for the tests to work we need to make a mirroring change to Canonical so that the tune text is displayed in the context pf any changed key (so that implicit pitches are treated properly).