Reworked Transposition
The code is now clearer.  Canonical reverts to its earlier simple version which just turns to text whichever AbcTune it is given without attempting to simplify otherwise irrelevant accidental markers which are already implied by the key signature.

Transposition has been reworked so that it contains the entire state for the transposition - i.e. source key, target key, source bar accidentals and target bar accidentals.  Two tests still fail - where a natural accidental has been placed in a bar and is followed by the sharpened or flattened note which is in the key signature.  In these cases the note needs to be explicitly sharpened or flattened, but it is left implicit and thus takes on the natural nature.