Added example producing MIDI output
The new midiplayer directory contains an example where the ABC is converted into the MidiRecording type (i.e. the same data type that is used in the MIDI file parser elm-comidi).  This should mean that eventually it will be possible to use the midi-player module with both sources.

In order to do this, more support for MIDI was added to Music.Notation. The data types used by Repeats were extracted and placed in their own file - RepeatTypes.  Finally, Repeat itself was modified so that Bars of music are polymorphic in note type (either MIDI using ticks or accumulated time using Floats).