Support degenerate 1st and 2nd repeats
When these are of the form [1 or [2 at the beginning of a line then they were rejected.  It's difficult to say from a close reading of the spec whether or not this is allowed, but there are plenty of tunes out in the wild which use them.

Normally you would expect |[1 and |[2 etc.