Got rid of NoteSequence
For some reason or other, I had allowed chords to appear within grace notes.  I can see absolutely no evidence for it - it must have been an aberration.  Grace notes now consist of an optional accacciatura and then just a list of notes.  This means we can now get rid of the awkward NoteSequence constructor for Music.  And all the code that catered for them in the parser, canonical, octave and transposition modules.